requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.4
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-jwt-auth==0.5.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
yarl==1.9.2
